Refactored getContactContent function into Contact object with proper methods.
Created the Contact object which extracts all the information about
a contact from a CalendarEvent object and can retrieve further
details through the contactId field if present.

The Contact object can generate the line of text with all the details
about the birthday of the interested contact both in HTML and in plain
text.

Added telephone numbers links (href="tel:xxxxxx").

Increased version number to 2.1.