Added notification for outdated versions and documented semantic versioning.
From now on users who are running an outdated version of the script
will see an additional line in the emails notifying it along with a link to the latest release.

Version check and comparison is performed via an implementation
of semantic versioning.

This solves the problem raised by @baatochan in #48.