Fix upstream SimplifiedSemanticVersion() generation
The object returned by the regex search of the downloaded script is an array
with matched subscripts from [1]...[n] - see:
 https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/exec
but function was trying to use whole object as a string, and was executing
inside a try-block, so always just returned negative (spilling error-line to
log). Fixed to use first match-subscript, i.e. latestVersion[1].