Prevent wrong contact names from being added to the email subject.
The subject was built using a list of all the contacts with events
on the given day, without considerng the `notification.eventTypes`
setting.
This meant that, for example, if a contact had an ANNIVERSARY event
on XX/YY/ZZZZ and the user set `notification.eventTypes` to filter
out anniversaries, while the email notification text did not display
the anniversary (correct) the notification subject included the name
of that contact (wrong).

Now the events excluded from the `notification.eventTypes` setting
are blacklisted and do not appear in the final list of contacts.

Closes #105.