Introduced a method to correctly add a number of days to a Date.
The previous approach (adding n * 24*60*60*1000 milliseconds to
the date timestamp) resulted in an incorrect date if the time
interval overlapped a DST change. This meant that the script could
fail in some edge-cases.

Added a function to the test suite to verify that the current
behavior of the Date functions (timezones offsets in particular)
do not change in the future.