Maximum length for email subject added.
Turns out the MailApp.sendEmail() API has an undocumented limit
of 250 characters on the length of the email subject. If the
subject is longer than the limit an error is thrown.

I've added a test to check if the limit is met: if it is not the
string gets truncated to 250 - 3 characaters and an ellipsis is
added at the end, so that the final string "ABCABCABC..." is 250
characters long.