add required alt attribute to img tag
HTML standard requires the IMG tag to contain an ALT attribute. As in this case the image is only for illustration I propose to supply an empty attribute to keep the code simple and the message small.

I assume people using screen readers or similar tools do use the text-version of the mail, so no need to specify something. 
It might be an option to use the full name again as well.