Fixed use of deprecated method Date.prototype.getYear().
The Google Script platform is activating support for the V8
JavaScript engine, abandoning the old Rhino interpreter.
In the old engine the Date.prototype.getYear() method returned the
full 4-digit year of a date, while in the new engine it returns
only the number of years since 1900.
This change led to issue #166.

The solution was simply to use the proper method:
Date.prototype.getFullYear().

See:
https://developers.google.com/apps-script/guides/v8-runtime
https://developers.google.com/apps-script/guides/v8-runtime/migration